+++ rex-engine/Cargo.toml
@@ -11,7 +11,6 @@ rand = { version = "0.9" }
 rex-ast = { path = "../rex-ast" }
 rex-lexer = { path = "../rex-lexer" }
 rex-parser = { path = "../rex-parser" }
-rex-proc-macro = { path = "../rex-proc-macro" }
 rex-type-system = { path = "../rex-type-system" }
 rpds = { version = "1.1" }
 thiserror = { version = "1.0" }
