+++ rex-engine/src/eval.rs
@@ -448,72 +448,17 @@ where
 #[cfg(test)]
 pub mod test {
     use rex_ast::{assert_expr_eq, b, d, f, i, l, n, s, tup, u};
-    use rex_lexer::Token;
-    use rex_parser::Parser;
-    use rex_proc_macro::Rex;
     use rex_type_system::{
         bool,
-        constraint::generate_constraints,
         dict, float, int, list, option, result, string,
-        trace::sprint_expr_with_type,
         tuple,
         types::Type,
         uint,
-        unify::{self},
-        types::ToType,
     };
 
-    use crate::engine::Builder;
+    use crate::util::parse_infer_and_eval;
 
     use super::*;
-
-    /// Helper function for parsing, inferring, and evaluating a given code
-    /// snippet. Pretty much all of the test suites can use this flow for
-    /// testing that the engine is correctly evaluating types and expressions.
-    /// In the future, we should probably make this (or some version of this) an
-    /// actual function for library users too.
-    async fn parse_infer_and_eval(code: &str) -> Result<(Expr, Type), Error> {
-        let builder: Builder<()> = Builder::with_prelude().unwrap();
-        parse_infer_and_eval_b(builder, code).await
-    }
-
-    /// This is similar to parse_infer_and_eval but lets you supply your own Builder,
-    /// in case you want to register any extra functions or ADTs.
-    async fn parse_infer_and_eval_b(
-        builder: Builder<()>,
-        code: &str,
-    ) -> Result<(Expr, Type), Error> {
-        let mut parser = Parser::new(Token::tokenize(code).unwrap());
-        let expr = parser.parse_expr().unwrap();
-
-        let (mut constraint_system, ftable, type_env) = builder.build();
-
-        let mut expr_type_env = ExprTypeEnv::new();
-        let ty = generate_constraints(&expr, &type_env, &mut expr_type_env, &mut constraint_system)
-            .unwrap();
-
-        let subst = unify::unify_constraints(&constraint_system).unwrap();
-        let res_type = unify::apply_subst(&ty, &subst);
-
-        println!(
-            "{}\n",
-            sprint_expr_with_type(&expr, &expr_type_env, Some(&subst))
-        );
-
-        let res = eval(
-            &Context {
-                scope: Scope::new_sync(),
-                ftable,
-                subst,
-                env: Arc::new(RwLock::new(expr_type_env)),
-                state: (),
-            },
-            &expr,
-        )
-        .await;
-
-        res.map(|res| (res, res_type))
-    }
 
     #[tokio::test]
     async fn test_literals() {
@@ -1429,158 +1374,5 @@ pub mod test {
             ];
             ignore span
         );
-    }
-
-    #[tokio::test]
-    async fn test_adt_enum() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub enum Color {
-            Red,
-            Green,
-            Blue,
-        }
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Color::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"(Red, Green, Blue)"#)
-            .await
-            .unwrap();
-        assert_eq!(
-            res_type,
-            tuple!(Color::to_type(), Color::to_type(), Color::to_type()));
-        assert_expr_eq!(
-            res,
-            tup!(n!("Red", None), n!("Green", None), n!("Blue", None));
-            ignore span);
-    }
-
-    #[tokio::test]
-    async fn test_adt_variant_tuple() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub enum Shape {
-            Rectangle(f64, f64),
-            Circle(f64),
-        }
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Shape::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"Rectangle (2.0 * 3.0) (4.0 * 5.0)"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, Shape::to_type());
-        assert_expr_eq!(
-            res,
-            n!("Rectangle", Some(tup!(f!(6.0), f!(20.0))));
-            ignore span);
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Shape::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"Circle (3.0 * 4.0)"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, Shape::to_type());
-        assert_expr_eq!(
-            res,
-            n!("Circle", Some(f!(12.0)));
-            ignore span);
-    }
-
-    #[tokio::test]
-    async fn test_adt_variant_struct() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub enum Shape {
-            Rectangle { width: f64, height: f64 },
-            Circle { radius: f64 },
-        }
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Shape::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"Rectangle { width = 2.0 * 3.0, height = 4.0 * 5.0 }"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, Shape::to_type());
-        assert_expr_eq!(
-            res,
-            n!("Rectangle", Some(d!(width = f!(6.0), height = f!(20.0))));
-            ignore span);
-    }
-
-    #[tokio::test]
-    async fn test_adt_struct() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub struct Movie {
-            pub title: String,
-            pub year: u16,
-        }
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Movie::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"Movie { title = "Godzilla", year = 1954 }"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, Movie::to_type());
-        assert_expr_eq!(
-            res,
-            n!("Movie", Some(d!(title = s!("Godzilla"), year = u!(1954))));
-            ignore span);
-    }
-
-    #[tokio::test]
-    async fn test_adt_tuple() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub struct Movie(pub String, pub u16);
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Movie::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"Movie "Godzilla" 1954 }"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, Movie::to_type());
-        assert_expr_eq!(
-            res,
-            n!("Movie", Some(tup!(s!("Godzilla"), u!(1954))));
-            ignore span);
-    }
-
-    #[tokio::test]
-    async fn test_adt_curry() {
-        #![allow(dead_code)]
-        #[derive(Rex)]
-        pub enum Shape {
-            Rectangle(f64, f64),
-            Circle(f64),
-        }
-
-        let mut builder: Builder<()> = Builder::with_prelude().unwrap();
-        builder.register_adt(&Shape::to_type()).unwrap();
-        let (res, res_type) = parse_infer_and_eval_b(
-            builder,
-            r#"let partial = Rectangle (2.0 * 3.0) in (partial (3.0 * 4.0), partial (2.0 * 4.0))"#)
-            .await
-            .unwrap();
-        assert_eq!(res_type, tuple!(Shape::to_type(), Shape::to_type()));
-        assert_expr_eq!(
-            res,
-            tup!(
-                n!("Rectangle", Some(tup!(f!(6.0), f!(12.0)))),
-                n!("Rectangle", Some(tup!(f!(6.0), f!(8.0)))));
-            ignore span);
     }
 }
