+++ rex-engine/src/lib.rs
@@ -3,3 +3,4 @@ pub mod engine;
 pub mod error;
 pub mod eval;
 pub mod ftable;
+pub mod util;
